Boxing/unboxing
git-svn-id: https://svn-us.apache.org/repos/asf/commons/proper/jexl/branches/2.0@803067 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/commons/jexl/examples/MethodPropertyTest.java
@@ -71,7 +71,7 @@ public class MethodPropertyTest extends TestCase {
 
         e = jexl.createExpression("foo.convert(1+number)");
         o = e.evaluate(jc);
-        out.print("value of " + e.getExpression() + " is : ", o, foo.convert(1+number));
+        out.print("value of " + e.getExpression() + " is : ", o, foo.convert(1+number.intValue()));
 
         /*
          * access a property
